chore(molecules): move status to src/components/molecules
