+++ app/ui/invoices/table.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image';
 import { UpdateInvoice, DeleteInvoice } from '@/src/components/molecules/buttons';
-import InvoiceStatus from '@/app/ui/invoices/status';
+import InvoiceStatus from '@/src/components/molecules/status';
 import { formatDateToLocal, formatCurrency } from '@/app/lib/utils';
 import { fetchFilteredInvoices } from '@/app/lib/data';
